server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.1.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.10.0
python-dotenv>=1.0.0

